refactor(Restart): 重构登录模块以支持更换庭院皮肤- 新增 LoginBase 类，实现庭院皮肤更换功能 - 修改 LoginHandler 类，集成 LoginBase 类功能 - 添加 harvest 主题的图像配置文件 - 优化登录流程，支持更换庭院皮肤后进行收获操作
